+++ examples/node/src/functionalities/placeBet.ts
@@ -7,19 +7,27 @@ import {
   type CasinoGame,
   type CasinoGameToken,
   type CasinoPlacedBet,
+  type CasinoRolledBet,
   type CasinoToken,
   type ChoiceInput,
   CoinToss,
   type CoinTossChoiceInput,
-  type CoinTossPlacedBet,
   Dice,
   type DiceChoiceInput,
-  type DicePlacedBet,
   FORMAT_TYPE,
   GAS_PRICE_TYPE,
+  Keno,
+  type KenoChoiceInput,
+  type NormalCasinoPlacedBet,
+  type NormalGameChoiceInput,
   Roulette,
   type RouletteChoiceInput,
-  type RoulettePlacedBet,
+  WEIGHTED_CASINO_GAME_TYPES,
+  type WeightedCasinoPlacedBet,
+  type WeightedGameChoiceInput,
+  type WeightedGameConfiguration,
+  Wheel,
+  type WheelChoiceInput,
   bigIntFormatter,
   casinoChains,
   chainById,
@@ -69,7 +77,15 @@ export async function startPlaceBetProcess() {
     const placedBet = await _placeBet(gameToken, selectedInput, betCount, betAmount);
 
     // 10. Wait for the roll
-    await _waitRoll(placedBet);
+    if (WEIGHTED_CASINO_GAME_TYPES.includes(placedBet.game)) {
+      await _waitWeightedRoll(
+        placedBet as WeightedCasinoPlacedBet,
+        (selectedInput as WeightedGameChoiceInput).config,
+        gameToken.affiliateHouseEdge,
+      );
+    } else {
+      await _waitRoll(placedBet as NormalCasinoPlacedBet);
+    }
   } catch (error) {
     if (error instanceof BetSwirlError) {
       console.error(
@@ -105,11 +121,13 @@ async function _selectGame(selectedChain: CasinoChain): Promise<CasinoGame> {
   const selectedGame = await select({
     message: "Select a game",
     loop: false,
-    choices: casinoGames.map((g) => ({
-      name: g.label,
-      value: g,
-      disabled: g.paused || [CASINO_GAME_TYPE.KENO].includes(g.game),
-    })),
+    choices: casinoGames
+      .map((g) => ({
+        name: g.label,
+        value: g,
+        disabled: g.paused,
+      }))
+      .filter((g) => g.value.game !== CASINO_GAME_TYPE.CUSTOM_WEIGHTED_GAME),
   });
   return selectedGame;
 }
@@ -178,8 +196,9 @@ async function _getTokenInfo(
 
 async function _selectInput(
   gameToken: CasinoGameToken,
-): Promise<CoinTossChoiceInput | DiceChoiceInput | RouletteChoiceInput> {
-  let input: CoinTossChoiceInput | DiceChoiceInput | RouletteChoiceInput;
+): Promise<NormalGameChoiceInput | WeightedGameChoiceInput> {
+  let input: NormalGameChoiceInput | WeightedGameChoiceInput;
+
   switch (gameToken.game) {
     case CASINO_GAME_TYPE.DICE:
       input = await select({
@@ -201,7 +220,7 @@ async function _selectInput(
         })),
       });
       break;
-    default: {
+    case CASINO_GAME_TYPE.ROULETTE: {
       const inputChoices = await checkbox({
         message: "Select a number or a bundle of numbers (space bar to select)",
         loop: false,
@@ -220,6 +239,47 @@ async function _selectInput(
           ),
         );
       }
+      break;
+    }
+    case CASINO_GAME_TYPE.KENO: {
+      // Get Keno config for the selected token
+      const kenoConfig = await wagmiBetSwirlClient.getKenoConfiguration(
+        gameToken,
+        gameToken.chainId,
+      );
+      input = await select({
+        message: "Select some balls",
+        loop: false,
+        choices: Keno.getChoiceInputs(kenoConfig, gameToken.affiliateHouseEdge).map((i) => ({
+          name: `${i.label} (${i.winChancePercent
+            .map(
+              (chance, index) =>
+                `${i.formattedNetMultiplier?.[index]}x - ${chance.toFixed(2)}% to win ${index !== i.winChancePercent.length - 1 ? "|" : ""}`,
+            )
+            .join("\n")})`,
+          value: i,
+        })),
+      });
+      break;
+    }
+    // Wheel
+    default: {
+      input = await select({
+        message: "Select a configuration",
+        loop: false,
+        // You could bring your own configurations here by passing customConfigurations in getChoiceInputs
+        choices: Wheel.getChoiceInputs(gameToken.chainId, gameToken.affiliateHouseEdge).map(
+          (i) => ({
+            name: `${i.label} (${i.winChancePercent
+              .map(
+                (chance, index) =>
+                  `${i.formattedNetMultiplier?.[index]}x - ${chance.toFixed(2)}% to win ${index !== i.winChancePercent.length - 1 ? "|" : ""}`,
+              )
+              .join("\n")})`,
+            value: i,
+          }),
+        ),
+      });
     }
   }
   return input;
@@ -345,10 +405,10 @@ async function _selectBetAmount(
 
 async function _placeBet(
   casinoGameToken: CasinoGameToken,
-  inputChoice: DiceChoiceInput | CoinTossChoiceInput | RouletteChoiceInput,
+  inputChoice: NormalGameChoiceInput | WeightedGameChoiceInput,
   betCount: number,
   betAmount: bigint,
-): Promise<CoinTossPlacedBet | DicePlacedBet | RoulettePlacedBet> {
+): Promise<CasinoPlacedBet> {
   const commonParams = {
     betCount,
     betAmount,
@@ -374,7 +434,8 @@ async function _placeBet(
   };
   let placedBetData: {
     receipt: TransactionReceipt;
-    placedBet: CoinTossPlacedBet | DicePlacedBet | RoulettePlacedBet;
+    placedBet: CasinoPlacedBet;
+    weightedGameConfig?: WeightedGameConfiguration;
   };
   if (inputChoice.game === CASINO_GAME_TYPE.DICE) {
     const diceCap = (inputChoice as DiceChoiceInput).value;
@@ -392,10 +453,28 @@ async function _placeBet(
       callbacks,
       casinoGameToken.chainId,
     );
-  } else {
+  } else if (inputChoice.game === CASINO_GAME_TYPE.ROULETTE) {
     const rouletteNumbers = (inputChoice as RouletteChoiceInput).value;
     placedBetData = await wagmiBetSwirlClient.playRoulette(
       { ...commonParams, numbers: rouletteNumbers },
+      undefined,
+      callbacks,
+      casinoGameToken.chainId,
+    );
+  } else if (inputChoice.game === CASINO_GAME_TYPE.KENO) {
+    const kenoChoice = inputChoice as KenoChoiceInput;
+    placedBetData = await wagmiBetSwirlClient.playKeno(
+      { ...commonParams, balls: kenoChoice.value, kenoConfig: kenoChoice.config },
+      undefined,
+      callbacks,
+      casinoGameToken.chainId,
+    );
+  }
+  // Wheel
+  else {
+    const weightedGameChoice = inputChoice as WheelChoiceInput;
+    placedBetData = await wagmiBetSwirlClient.playWheel(
+      { ...commonParams, weightedGameConfig: weightedGameChoice.config },
       undefined,
       callbacks,
       casinoGameToken.chainId,
@@ -414,7 +493,7 @@ async function _placeBet(
   return placedBetData.placedBet;
 }
 
-async function _waitRoll(placedBet: CasinoPlacedBet) {
+async function _waitRoll(placedBet: NormalCasinoPlacedBet) {
   console.log(chalk.blue("⌛ Waiting the bet to be rolled..."));
   const rolledBetData = await wagmiBetSwirlClient.waitRolledBet(placedBet, {
     timeout: 300000, //5min
@@ -423,6 +502,31 @@ async function _waitRoll(placedBet: CasinoPlacedBet) {
   });
 
   const rolledBet = rolledBetData.rolledBet;
+  _displayRolledBet(rolledBet);
+}
+
+async function _waitWeightedRoll(
+  placedBet: WeightedCasinoPlacedBet,
+  weightedGameConfig: WeightedGameConfiguration,
+  houseEdge: number,
+) {
+  console.log(chalk.blue("⌛ Waiting the bet to be rolled..."));
+  const rolledBetData = await wagmiBetSwirlClient.waitRolledBet(
+    placedBet,
+    {
+      timeout: 300000, //5min
+      pollingInterval: process.env.RPC_URL ? 500 : 2500,
+      formatType: FORMAT_TYPE.FULL_PRECISE,
+    },
+    weightedGameConfig,
+    houseEdge,
+  );
+
+  const rolledBet = rolledBetData.rolledBet;
+  _displayRolledBet(rolledBet);
+}
+
+function _displayRolledBet(rolledBet: CasinoRolledBet) {
   const chain = chainById[rolledBet.chainId];
   const commonMessage = chalk.blue(
     `Payout: ${rolledBet.formattedPayout} ${
@@ -436,7 +540,7 @@ async function _waitRoll(placedBet: CasinoPlacedBet) {
     )}\nRoll txn: ${formatTxnUrl(rolledBet.rollTxnHash, rolledBet.chainId)}\nBetSwirl url: ${getBetSwirlBetUrl(rolledBet.id, rolledBet.game, rolledBet.chainId)}`,
   );
   // Win
-  if (rolledBetData.rolledBet.isWin) {
+  if (rolledBet.isWin) {
     console.log(
       chalk.green(
         `🥳 Congrats you won ${rolledBet.formattedBenefit} ${rolledBet.token.symbol} (x${rolledBet.formattedPayoutMultiplier})\n`,