+++ examples/node/src/functionalities/showHistoryBets.ts
@@ -6,6 +6,8 @@ import {
   type CasinoChain,
   FORMAT_TYPE,
   OrderDirection,
+  WEIGHTED_CASINO_GAME_TYPES,
+  WeightedGame,
   bigIntFormatter,
   casinoChains,
   formatTxnUrl,
@@ -18,7 +20,7 @@ import chalk from "chalk";
 import type { Hex } from "viem/_types/types/misc";
 import { checkEnvVariables, getWagmiConfigFromCasinoChain } from "../../utils";
 
-let wagmiBetSwielClient: WagmiBetSwirlClient;
+let wagmiBetSwirlClient: WagmiBetSwirlClient;
 
 export async function startShowHistoryBetsProcess() {
   try {
@@ -55,7 +57,7 @@ async function _selectChain(): Promise<CasinoChain> {
     choices: casinoChains.map((c) => ({ name: c.viemChain.name, value: c })),
   });
   const wagmiConfig = getWagmiConfigFromCasinoChain(selectedChain);
-  wagmiBetSwielClient = initWagmiBetSwirlClient(wagmiConfig, {
+  wagmiBetSwirlClient = initWagmiBetSwirlClient(wagmiConfig, {
     chainId: selectedChain.id,
     affiliate: process.env.AFFILIATE_ADDRESS as Hex,
     formatType: FORMAT_TYPE.PRECISE,
@@ -68,10 +70,10 @@ async function _selectChain(): Promise<CasinoChain> {
 }
 
 async function _getLastBets(count: number, casinoChain: CasinoChain): Promise<CasinoBet[]> {
-  const { bets, error } = await wagmiBetSwielClient.fetchBets(
+  const { bets, error } = await wagmiBetSwirlClient.fetchBets(
     casinoChain.id,
     {
-      bettor: wagmiBetSwielClient.betSwirlWallet.getAccount()?.address,
+      bettor: wagmiBetSwirlClient.betSwirlWallet.getAccount()?.address,
     },
     1,
     count,
@@ -102,7 +104,9 @@ function _showBet(bet: CasinoBet) {
   // Common place bet info
   // TODO replace "Input" and "Rolled" by the game input/output labels
   const placeBetInfo = `Game: ${labelCasinoGameByType[bet.game]}\nInput: ${
-    bet.decodedInput
+    WEIGHTED_CASINO_GAME_TYPES.includes(bet.game)
+      ? WeightedGame.getWeightedGameConfigLabel(bet.decodedInput, bet.chainId)
+      : bet.decodedInput
   }\nBet amount: ${bet.formattedBetAmount} ${bet.token.symbol}\nBet count: ${bet.betCount}\n${
     bet.betCount > 1 ? `Total bet amount ${bet.formattedTotalBetAmount} ${bet.token.symbol}\n` : ""
   }Bet date: ${bet.betDate.toLocaleString()}\nBet txn: ${formatTxnUrl(bet.betTxnHash, bet.chainId)}\n`;