+++ packages/core/src/actions/casino/game.ts
@@ -1,15 +1,18 @@
 import { type Hash, type Hex, type TransactionReceipt, encodeFunctionData } from "viem";
 import { decodeEventLog } from "viem";
-import { coinTossAbi } from "../../abis/v2/casino/coinToss";
+import { coinTossAbi } from "../../abis/v2/casino/cointoss";
 import { GAS_TOKEN_ADDRESS } from "../../constants";
 import {
   CASINO_GAME_TYPE,
   type CasinoChain,
   type CasinoChainId,
   MAX_SDK_HOUSE_EGDE,
+  type NORMAL_CASINO_GAME_TYPE,
+  type WEIGHTED_CASINO_GAME_TYPE,
   casinoChainById,
 } from "../../data/casino";
-import type { CoinTossEncodedInput } from "../../entities/casino/coinToss";
+import type { KenoEncodedInput } from "../../entities";
+import type { CoinTossEncodedInput } from "../../entities/casino/cointoss";
 import type { DiceEncodedInput } from "../../entities/casino/dice";
 import type { RouletteEncodedInput } from "../../entities/casino/roulette";
 import { ERROR_CODES } from "../../errors/codes";
@@ -56,14 +59,17 @@ export const defaultCasinoPlaceBetOptions = {
   allowanceType: ALLOWANCE_TYPE.AUTO,
 };
 // Game should not know the game implementation details, but well..  it helps developers
-export type GameEncodedInput = CoinTossEncodedInput | DiceEncodedInput | RouletteEncodedInput;
-
+export type GameEncodedInput =
+  | CoinTossEncodedInput
+  | DiceEncodedInput
+  | RouletteEncodedInput
+  | KenoEncodedInput;
 export interface GenericCasinoBetParams extends CasinoBetParams {
   game: CASINO_GAME_TYPE;
   gameEncodedInput: GameEncodedInput;
 }
 
-export interface CasinoPlacedBet {
+interface CommonCasinoPlacedBet {
   id: bigint;
   betAmount: bigint;
   betCount: number;
@@ -77,8 +83,15 @@ export interface CasinoPlacedBet {
   betTxnHash: Hash;
   betBlock: bigint;
   chainId: CasinoChainId;
-  game: CASINO_GAME_TYPE;
 }
+export interface NormalCasinoPlacedBet extends CommonCasinoPlacedBet {
+  game: NORMAL_CASINO_GAME_TYPE;
+}
+export interface WeightedCasinoPlacedBet extends CommonCasinoPlacedBet {
+  game: WEIGHTED_CASINO_GAME_TYPE;
+}
+
+export type CasinoPlacedBet = NormalCasinoPlacedBet | WeightedCasinoPlacedBet;
 
 export interface PlaceBetCallbacks {
   onApprovePending?: (tx: Hash, result: ApproveResult) => void | Promise<void>;