+++ packages/core/src/read/casino/keno.ts
@@ -1,14 +1,39 @@
-import { type Address, type Hex, encodeFunctionData } from "viem";
+import { type Address, type Hex, type TransactionReceipt, encodeFunctionData } from "viem";
 import { kenoAbi } from "../../abis";
+import type { KenoPlacedBet } from "../../actions/casino/keno";
 import { CASINO_GAME_TYPE, type CasinoChainId, casinoChainById } from "../../data/casino";
+import { Keno, type KenoBall } from "../../entities/casino/keno";
 import { ERROR_CODES, TransactionError } from "../../errors";
 import { ChainError } from "../../errors";
 import type { BetSwirlFunctionData, Token } from "../../interfaces";
 import type { BetSwirlWallet } from "../../provider";
 import { getCasinoChainId } from "../../utils";
+import type { CasinoRolledBet, CasinoWaitRollOptions } from "./game";
+
+export interface KenoRolledBet extends Omit<CasinoRolledBet, "decodedRoll"> {
+  rolled: KenoBall[][];
+}
+
+export async function waitKenoRolledBet(
+  wallet: BetSwirlWallet,
+  placedBet: KenoPlacedBet,
+  options?: CasinoWaitRollOptions,
+): Promise<{
+  rolledBet: KenoRolledBet;
+  receipt: TransactionReceipt;
+}> {
+  const { rolledBet, receipt } = await waitKenoRolledBet(wallet, placedBet, options);
+  return {
+    rolledBet: {
+      ...rolledBet,
+      rolled: rolledBet.encodedRolled.map(Keno.decodeRolled),
+    },
+    receipt,
+  };
+}
 
 /**
- * Raw token info data returned by the smart contract
+ * Raw Keno configuration data returned by the smart contract
  * [0] - biggestNumber: The biggest selectable number
  * [1] - maxNumbersPlayed: Maximum selectable numbers
  * [2] - gainsTable: The gain multipliers (gain multiplier = gains[numbers played][numbers matched]) (BP)
@@ -23,17 +48,20 @@ export function parseRawKenoConfiguration(
   return {
     token,
     chainId: casinoChainId,
-    biggestSelectableNumber: Number(rawConfiguration[0]),
-    maxSelectableNumbers: Number(rawConfiguration[1]),
-    mutliplierTable: rawConfiguration[2].map((row) => row.map((multiplier) => Number(multiplier))),
+    biggestSelectableBall: Number(rawConfiguration[0]),
+    maxSelectableBalls: Number(rawConfiguration[1]),
+    mutliplierTable: [
+      [0],
+      ...rawConfiguration[2].map((row) => row.map((multiplier) => Number(multiplier))),
+    ],
   };
 }
 
 export interface KenoConfiguration {
   token: Token;
   chainId: CasinoChainId;
-  biggestSelectableNumber: number;
-  maxSelectableNumbers: number;
+  biggestSelectableBall: number;
+  maxSelectableBalls: number;
   mutliplierTable: number[][]; // BP
 }