+++ packages/core/src/read/index.ts
@@ -1,10 +1,12 @@
 /* casino */
 export * from "./casino/bank";
-export * from "./casino/coinToss";
+export * from "./casino/cointoss";
 export * from "./casino/dice";
 export * from "./casino/roulette";
 export * from "./casino/keno";
 export * from "./casino/game";
+export * from "./casino/weightedGame";
+export * from "./casino/wheel";
 
 /* common */
 export * from "./common/chainlinkVrfCost";