+++ packages/providers/wagmi/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @betswirl/wagmi-provider
 
+## 0.1.0
+
+### Minor Changes
+
+- Add Wheel game with weighted gamr logic
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @betswirl/sdk-core@0.1.0
+
+## 0.0.9
+
+### Patch Changes
+
+- Add Keno Game
+
+- Updated dependencies []:
+  - @betswirl/sdk-core@0.0.9
+
 ## 0.0.8
 
 ### Patch Changes